Build multi-platform Docker images/manifests #864
The goal of this change is to make it trivial to run
farmOS on the most common ARM architectures without the
need for a local `docker build` step.

Unfortunately, the change is fairly involved in order to keep
the caching and testing working the same way.

Ideally, we'd be able to do a single `docker buildx build --push --platform COMMA_SEPARATED_PLATFORMS_HERE --tag`
command, but that wouldn't work for how we test pull requests and handle
creating the various tags conditionally - in addition to making sure our
test workflows run against the same images prior to actually pushing
them to Docker hub.

Instead, we have to do the more manual strategy of using `docker buildx`
to build/cache each platform's image individually, then only at the end
do we create the final multi-architecture manifest.

References:

* https://www.docker.com/blog/multi-arch-build-and-images-the-simple-way/
* https://github.com/docker/setup-qemu-action
* https://github.com/docker/setup-buildx-action